Add scheduler driver. Uses a timer to schedule function calls after a timeout
The following values must be defined:
SCHED_TIMER: Timer module to use (0 = TIMER_A, 1 = TIMER_B)
SCHED_DIV:   Amount to divide SMCLK by to give a 1kHz interrupt
SCHED_QUEUE_SIZE: The size of the internal schedule queue

If you attempt to add more tasks to the queue than 'SCHED_QUEUE_SIZE' then
the code will get stuck in a while(1) loop.

Add tasks with the sched_add(t) function where t is a struct of type
sched_task_t containing a pointer to the callback and a time to wait.

Remove a task by passing a pointer to the sched_task_t struct of the task
to sched_rem(t).

If a callback returns true then it will be re-scheduled with the same
timeout as before. Otherwise it will be removed from the queue.