Disable *all* interrupts when modifying the scheduler queue
The sched_add function modifies the scheduler queue.  It can be called
from an interrupt context, or that other context that remains nameless
(uh, the "mainloop").

Previously, the sched_add function would disable only its own timer
interrupt.  However, this is not sufficient, as sched_add should be
callable from other interrupt routines.  If it were to be called from,
say, the pin change ISR, then unintended consequences could result.

The sched_add function would iterate through the queue looking for an
empty slot.  When it found one, it would *then* disable interrupts.
If an interrupt were to happen between the emptiness comparison
(task==NULL) and disabling interrupts, and call sched_add, then this
slot would no longer be empty.  The task added by the interrupt
routine would be immediately overwritten, which is not good.

This patch corrects both of these problems by checking that the task
slot is empty after disabling *all* interrupts.  If it isn't then it
re-enables interrupts and continues through the queue.  It also marks
the scheduler queue as volatile so that this comparison will work as
desired.  (Also makes the queue static since it's not accessible from
outside sched.o)